Systemd services: obc-p1.service, start_gpsd.service

+++ sdr-testing/filters.h
@@ -8,9 +8,10 @@
 
 #define IGNORE_SAMPLES 2
 
+
 typedef struct
 {
-    int16_t buf[17];
+    int32_t buf[17];
     int16_t head;
     int16_t tail; 
     int16_t wait;
@@ -18,9 +19,10 @@ typedef struct
     int curr_s; // Current sample, increments until buffer size, then curr_s -= 2. Apply filter when curr_s == SIZE_16-1
 } filter_16;
 
+
 typedef struct 
 {
-    int16_t buf[33];
+    int32_t buf[33];
     int16_t head;
     int16_t tail; 
     int16_t wait;
@@ -28,9 +30,10 @@ typedef struct
     int curr_s; // Current sample, incremetns until buffer size, then curr_s -= 2. Apply filter when curr_s == SIZE_32-1
 } filter_32;
 
+
 typedef struct 
 {
-    int16_t buf[129];
+    int32_t buf[129];
     int16_t head;
     int16_t tail;
     int16_t wait;
@@ -38,14 +41,17 @@ typedef struct
     int curr_s; // Current sample, increments until buffer size, then curr_s -= 2. Apply filter when curr_s == SIZE_128-1
 } filter_128;
 
+// fir(16, 0.25) * 1024
 int16_t fir_16[17] = {
     0, -4, -12, -17, 0, 55, 140, 221, 255, 221, 140, 55, 0, -17, -12, -4, 0
 };
 
+// fir(32, 0.38) * 1024
 int16_t fir_32[33] = { 
     0, -2, -2, 1, 6, 5, -7, -16, -3, 26, 30, -16, -70, -43, 108, 301, 390, 301, 108, -43, -7, -16 , 30 , 26 , -3, -16, -7, 5, 6, 1, -2, -2, 0
 };
 
+// fir(128, 0.47) * 1024
 int16_t fir_128[129] = { 
     0 , 0, 0, 0, 0, 0, 0, 0, 1, 0, -1, 0, 1 , 0, -1, 0,  1, 0, -2,-1,  2,  1,- 2, -2,  2,  3, -1, -4,  1,   4,  0,  -5,
     -1, 6, 2,-6,-4, 7, 5,-6,-7, 6, 10,-5,-12, 3, 15,-1,-18,-3, 20, 8,-23,-14, 25, 23,-27,-36, 29, 57,-30,-104, 30, 324, 
@@ -53,7 +59,7 @@ int16_t fir_128[129] = {
     6,-1,-5, 0, 4, 1,-4,-1, 3, 2,-2,-2, 1, 2,-1,-2, 0, 1, 0,-1, 0, 1, 0,-1, 0, 1, 0, 0, 0, 0, 0, 0, 0,0
 };
 
-int16_t mult_16(filter_16 filter){
+int32_t mult_16(filter_16 filter){
     int32_t aux = 0;
     for( int i = 1; i < 16; i++){
         aux += filter.buf[i]*fir_16[i];
@@ -61,7 +67,7 @@ int16_t mult_16(filter_16 filter){
     return aux;
 }
 
-int16_t mult_32(filter_32 filter){
+int32_t mult_32(filter_32 filter){
     int32_t aux = 0;
     for( int i = 1; i < 32; i++){
         aux += filter.buf[i]*fir_32[i];
@@ -69,7 +75,7 @@ int16_t mult_32(filter_32 filter){
     return aux;
 }
 
-int16_t mult_128(filter_128 filter){
+int32_t mult_128(filter_128 filter){
     int32_t aux = 0;
     for( int i = 7; i < 121; i++){
         aux += filter.buf[i]*fir_128[i];
